freetype: add libpng to freetype-config and freetype2.pc
Add libpng libraries to freetype-config and freetype2.pc when they're
needed to avoid build breakage for other packages.

Patch in a different form is already upstream.

Fixing configure & friends is not so good since autoreconf busts things
up because of the odd way things are done (upstream fixed too it seems).

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>