qt5webkit: disable for MIPS64
qt5webkit is not currently supported on MIPS64 platforms, so disable the
possibility of selecting this package in that platform.

In the future the following changes would be needed to fix this problem:
A new Source/JavaScriptCore/assembler/MacroAssemblerMIPS64.h file needs
to be created in order to add support for MIPS64 and MacroAssembler.h
needs to be modified to include that file in case of CPU(MIPS64).
Also Source/WTF/wtf/Platform.h and Source/WTF/wtf/dtoa/utils.h need to
be modified to add MIPS64 support.

Signed-off-by: Vicente Olivert Riera <Vincent.Riera@imgtec.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>