package/python-pycrypto: fix legal-info
The pycrypto package contains a LICENSE.orig file, but apply-patches.sh
removes all .orig file, so we can't list a .orig file asa license file,
as it would have disapeared by the time 'make legal-info' sould be called.

Add a post-extract hook that renames this license file, so it is renamed
just before we apply patches and remove .orig files.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Cc: Ryan Barnett <rjbarnet@rockwellcollins.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>