Config: default BR2_PACKAGE_OVERRIDE_FILE to $(CONFIG_DIR)/local.mk
The current default for BR2_PACKAGE_OVERRIDE_FILE points to:
    $(TOPDIR)/local.mk

This works well for in-tree builds, but is not very useful for
out-of-tree builds, when the Buildroot source tree may be shared for
different concurrent builds.

Also, it seems to be more sensible to have local.mk alognside
the .config file.

Hence, change the default for BR2_PACKAGE_OVERRIDE_FILE to point to:
    $(CONFIG_DIR)/local.mk

Note that this does not change the current behaviour for in-tree
builds, since in that case $(CONFIG_DIR) == $(TOPDIR).

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Thomas De Schampheleire <patrickdepinguin@gmail.com>
Acked-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>