board/arm/foundation-v8: update defconfig and instructions
This commit updates the defconfig of the foundation-v8 platform, which
can be tested under the AArch64 emulator provided by ARM. The main
change is that we switch to the mainline 3.13.6 kernel instead of a
Linaro-specific Git tree.

Another change is to update the download link for the emulator in the
instructions, in order to use the latest version of the emulator,
which was used for testing.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>