package/x11-video-drivers: do not select, but depends on mesa3d
Currently, X.Org drivers that need mesa3d forcibly select it.

We'll soon switch mesa3d to being a provider of the full OpenGL virtual
package, to come in a later patch.

It is bad practice to select a provider, as it can generate configuration
inconsistencies.

So, switch all X.Org video driver that need mesa3d to actually depend on
it, rather than forcibly select it.

To be noted: xf86-video-mach64 already used a 'depends on' rather
than a 'select'.

At the same time, move the intel drivers dependencies to the top.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Bernd Kuhls <bernd.kuhls@t-online.de>
Cc: Paul Cercueil <paul@crapouillou.net>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>