libwebsockets: disable checks for C++ compiler
Since libwesockets in written in pure C, no C++, there is no reason to
disable it or make the configuration failed when the toolchain has no
C++ support.

Unfortunately, CMake defaults sets C and C++ as languages used by the
project. This makes the configuration failed if no C++ compiler is found.

This patch just make the CMake code of libwebsockets reflecting the real
status of the project with regard to the languages used.

Should fix:
  http://autobuild.buildroot.org/results/fef/fef98ef289d44d13ec61025c50febba1e835b421/

Signed-off-by: Samuel Martin <s.martin49@gmail.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>