configs/raspberrypi: bump kernel version
Upstream has been playing evil-rebase on their tree, so they now
no longer have the changeset we are currently referencing.

They have now agreed to merge instead of rebasing:
    https://github.com/raspberrypi/linux/issues/584

So, update the raspberrypi_defconfig to use a new changeset that
currently exists right _now_, in the hope it will not disapear in
the future (but we'll catch it pretty fast anyway, so we'll be able
to seriously bug up upstream in that case ;-) ).

Reported-by: Jeremy Rosen <jeremy.rosen@openwide.fr>
Signed-off-by: yann.morin.1998@free.fr
Cc: Jeremy Rosen <jeremy.rosen@openwide.fr>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>