package/tzdata: only compile the zoneinfo once
We already compile the zoneinfo once for the host, so there is
no need to recompile it for the target, we can just re-use the
host files.

This has no impact on tz (the uClibc counterpart of tzdata), since
it already made use of the host files anyway.

This means we no longer need to extract the target variant of
tzdata, too.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>